#30: Added an enum for missing state.
Since HomeVC will be displaying a brand new state, we need to expand our horizons. A `Bool` will only allow us to express granted or denied. Our new enum helps with undefined.

We could've used a `Bool?` and that would've added a third state, however an enum is more readable.